fix: Allow RDS for MySQL to connect via TLS (#122)
