+++ src/rds/triggers/util.ts
@@ -1,3 +1,4 @@
+import * as https from "https";
 // eslint-disable-next-line import/no-extraneous-dependencies
 import {
   SecretsManagerClient,
@@ -36,4 +37,26 @@ export async function fetchAllSecrets<T = Record<string, any>>(
     return [];
   }
   return Promise.all(arns.map((a) => fetchSecret<T>(a, client)));
+}
+
+/**
+ * Reads an HTTPS resource into a string.
+ *
+ * We need this function since newer RDS CA certificates aren't in Lambda.
+ *
+ * @see https://github.com/aws/aws-lambda-base-images/issues/123
+ */
+export function readRemote(
+  url: string,
+  options?: https.RequestOptions
+): Promise<Buffer> {
+  return new Promise((resolve, reject) => {
+    https
+      .get(url, options || {}, (res) => {
+        const data: Buffer[] = [];
+        res.on("data", (d) => data.push(d));
+        res.on("end", () => resolve(Buffer.concat(data)));
+      })
+      .on("error", (e) => reject(e));
+  });
 }
